# Edzkool/backend/users/openapi.yaml
openapi: 3.0.3
info:
  title: Users API (minimal)
  version: '1.0'
  description: |
    OpenAPI specification describing the current Users-related endpoints
    in this project. It includes the minimal CRUD `users` endpoints,
    a standalone `users/by-email` endpoint, OTP endpoints under `/otp/`,
    and a legacy `/send-otp/` endpoint.
servers:
  - url: http://127.0.0.1:8000/api/v1
paths:
...
  /otp/create/:
    post:
      summary: Create an OTP and send by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPCreate'
      responses:
        '201':
          description: OTP created and email attempted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPResponse'
  /otp/verify/:
    post:
      summary: Verify an OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid OTP
...
components:
  schemas:
...
    OTPCreate:
      type: object
      required:
        - otp_type
        - contact
      properties:
        otp_type:
          type: string
          description: 'Type of OTP (register, password_reset, etc.)'
        contact:
          type: string
          description: 'Email address to send OTP to'
    OTPVerifyRequest:
      type: object
      required:
        - otp_code
        - contact
      properties:
        otp_code:
          type: string
        contact:
          type: string
...
    OTPResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...